+++ src/wasm/wasm-type.cpp
@@ -62,7 +62,9 @@ std::vector<std::unique_ptr<std::vector<Type>>> typeLists = [] {
   add({Type::f32});
   add({Type::f64});
   add({Type::v128});
+  add({Type::funcref});
   add({Type::anyref});
+  add({Type::nullref});
   add({Type::exnref});
   return lists;
 }();
@@ -75,7 +77,9 @@ std::unordered_map<std::vector<Type>, uint32_t> indices = {
   {{Type::f32}, Type::f32},
   {{Type::f64}, Type::f64},
   {{Type::v128}, Type::v128},
+  {{Type::funcref}, Type::funcref},
   {{Type::anyref}, Type::anyref},
+  {{Type::nullref}, Type::nullref},
   {{Type::exnref}, Type::exnref},
 };
 
@@ -140,14 +144,122 @@ bool Type::operator<(const Type& other) const {
     [](const Type& a, const Type& b) { return uint32_t(a) < uint32_t(b); });
 }
 
-bool Signature::operator<(const Signature& other) const {
-  if (results < other.results) {
-    return true;
-  } else if (other.results < results) {
-    return false;
-  } else {
-    return params < other.params;
+unsigned Type::getByteSize() const {
+  assert(isSingle() && "getByteSize does not works with single types");
+  Type singleType = *expand().begin();
+  switch (singleType) {
+    case Type::i32:
+      return 4;
+    case Type::i64:
+      return 8;
+    case Type::f32:
+      return 4;
+    case Type::f64:
+      return 8;
+    case Type::v128:
+      return 16;
+    case Type::funcref:
+    case Type::anyref:
+    case Type::nullref:
+    case Type::exnref:
+    case Type::none:
+    case Type::unreachable:
+      WASM_UNREACHABLE("invalid type");
   }
+  WASM_UNREACHABLE("invalid type");
+}
+
+Type Type::reinterpret() const {
+  assert(isSingle() && "reinterpretType only works with single types");
+  Type singleType = *expand().begin();
+  switch (singleType) {
+    case Type::i32:
+      return f32;
+    case Type::i64:
+      return f64;
+    case Type::f32:
+      return i32;
+    case Type::f64:
+      return i64;
+    case Type::v128:
+    case Type::funcref:
+    case Type::anyref:
+    case Type::nullref:
+    case Type::exnref:
+    case Type::none:
+    case Type::unreachable:
+      WASM_UNREACHABLE("invalid type");
+  }
+  WASM_UNREACHABLE("invalid type");
+}
+
+FeatureSet Type::getFeatures() const {
+  FeatureSet feats = FeatureSet::MVP;
+  for (Type t : expand()) {
+    switch (t) {
+      case Type::v128:
+        feats |= FeatureSet::SIMD;
+        break;
+      case Type::anyref:
+        feats |= FeatureSet::ReferenceTypes;
+        break;
+      case Type::exnref:
+        feats |= FeatureSet::ExceptionHandling;
+        break;
+      default:
+        break;
+    }
+  }
+  return feats;
+}
+
+Type Type::get(unsigned byteSize, bool float_) {
+  if (byteSize < 4) {
+    return Type::i32;
+  }
+  if (byteSize == 4) {
+    return float_ ? Type::f32 : Type::i32;
+  }
+  if (byteSize == 8) {
+    return float_ ? Type::f64 : Type::i64;
+  }
+  if (byteSize == 16) {
+    return Type::v128;
+  }
+  WASM_UNREACHABLE("invalid size");
+}
+
+bool Type::Type::isSubType(Type left, Type right) {
+  if (left == right) {
+    return true;
+  }
+  if (left.isRef() && right.isRef() &&
+      (right == Type::anyref || left == Type::nullref)) {
+    return true;
+  }
+  return false;
+}
+
+Type Type::Type::getLeastUpperBound(Type a, Type b) {
+  if (a == b) {
+    return a;
+  }
+  if (a == Type::unreachable) {
+    return b;
+  }
+  if (b == Type::unreachable) {
+    return a;
+  }
+  if (!a.isRef() || !b.isRef()) {
+    return none; // a poison value that must not be consumed
+  }
+  if (a == Type::nullref) {
+    return b;
+  }
+  if (b == Type::nullref) {
+    return a;
+  }
+  return Type::anyref;
 }
 
 namespace {
@@ -169,6 +281,22 @@ template<typename T> std::string genericToString(const T& t) {
 }
 
 } // anonymous namespace
+
+std::string Type::toString() const { return genericToString(*this); }
+
+std::string ParamType::toString() const { return genericToString(*this); }
+
+std::string ResultType::toString() const { return genericToString(*this); }
+
+bool Signature::operator<(const Signature& other) const {
+  if (results < other.results) {
+    return true;
+  } else if (other.results < results) {
+    return false;
+  } else {
+    return params < other.params;
+  }
+}
 
 std::ostream& operator<<(std::ostream& os, Type type) {
   switch (type) {
@@ -193,8 +321,14 @@ std::ostream& operator<<(std::ostream& os, Type type) {
     case Type::v128:
       os << "v128";
       break;
+    case Type::funcref:
+      os << "funcref";
+      break;
     case Type::anyref:
       os << "anyref";
+      break;
+    case Type::nullref:
+      os << "nullref";
       break;
     case Type::exnref:
       os << "exnref";
@@ -224,89 +358,6 @@ std::ostream& operator<<(std::ostream& os, ResultType param) {
 
 std::ostream& operator<<(std::ostream& os, Signature sig) {
   return os << "Signature(" << sig.params << " => " << sig.results << ")";
-}
-
-std::string Type::toString() const { return genericToString(*this); }
-
-std::string ParamType::toString() const { return genericToString(*this); }
-
-std::string ResultType::toString() const { return genericToString(*this); }
-
-unsigned getTypeSize(Type type) {
-  switch (type) {
-    case Type::i32:
-      return 4;
-    case Type::i64:
-      return 8;
-    case Type::f32:
-      return 4;
-    case Type::f64:
-      return 8;
-    case Type::v128:
-      return 16;
-    case Type::anyref: // anyref type is opaque
-    case Type::exnref: // exnref type is opaque
-    case Type::none:
-    case Type::unreachable:
-      WASM_UNREACHABLE("invalid type");
-  }
-  WASM_UNREACHABLE("invalid type");
-}
-
-FeatureSet getFeatures(Type type) {
-  FeatureSet feats = FeatureSet::MVP;
-  for (Type t : type.expand()) {
-    switch (t) {
-      case v128:
-        feats |= FeatureSet::SIMD;
-        break;
-      case anyref:
-        feats |= FeatureSet::ReferenceTypes;
-        break;
-      case exnref:
-        feats |= FeatureSet::ExceptionHandling;
-        break;
-      default:
-        break;
-    }
-  }
-  return feats;
-}
-
-Type getType(unsigned size, bool float_) {
-  if (size < 4) {
-    return Type::i32;
-  }
-  if (size == 4) {
-    return float_ ? Type::f32 : Type::i32;
-  }
-  if (size == 8) {
-    return float_ ? Type::f64 : Type::i64;
-  }
-  if (size == 16) {
-    return Type::v128;
-  }
-  WASM_UNREACHABLE("invalid size");
-}
-
-Type reinterpretType(Type type) {
-  switch (type) {
-    case Type::i32:
-      return f32;
-    case Type::i64:
-      return f64;
-    case Type::f32:
-      return i32;
-    case Type::f64:
-      return i64;
-    case Type::v128:
-    case Type::anyref:
-    case Type::exnref:
-    case Type::none:
-    case Type::unreachable:
-      WASM_UNREACHABLE("invalid type");
-  }
-  WASM_UNREACHABLE("invalid type");
 }
 
 } // namespace wasm